@@ -15,7 +15,7 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 27
+    compileSdkVersion 31
     defaultConfig {
         applicationId "com.google.ar.sceneform.samples.hellosceneform"
 
@@ -59,15 +59,20 @@ android {
 dependencies {
     implementation files('src/main/libs/Java-WebSocket-1.3.8.jar')
     implementation files('src/main/libs/json-20210307.jar')
-    compile fileTree(include: ['*.jar'], dir: 'libs')
-    implementation 'com.google.ar.sceneform:core:1.4.0'
-    implementation 'com.google.ar.sceneform.ux:sceneform-ux:1.4.0'
-    implementation 'com.android.support:appcompat-v7:27.1.1'
-    implementation 'com.android.support:design:27.1.1'
-    implementation 'com.android.support.constraint:constraint-layout:1.1.3'
+    implementation fileTree(include: ['*.jar'], dir: 'libs')
+//    implementation 'com.google.ar.sceneform:core:1.4.0'
+    implementation 'com.gorisse.thomas.sceneform:core:1.21.0'
+//    implementation 'com.google.ar.sceneform.ux:sceneform-ux:1.17.1'
+    implementation 'com.gorisse.thomas.sceneform:ux:1.21.0'
+//    implementation 'com.android.support:appcompat-v7:28.0.0'
+    implementation 'androidx.appcompat:appcompat:1.5.1'
+//    implementation 'com.android.support:design:28.0.0'
+//    implementation 'com.android.support.constraint:constraint-layout:2.0.4'
+    implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
 //    compile files('app/src/main/libs/Java-WebSocket-1.3.8.jar')
 //    compile files('app/src/main/libs/json-20210307.jar')
     implementation project(':ImageProcessingLib')
+    implementation 'androidx.core:core:1.8.0'
 }
 
 
